refactor(constants): group all String and integers into one package
