+++ src/main/java/vendingmachine/constants/ExceptionMessages.java
@@ -1,4 +1,4 @@
-package vendingmachine.utils;
+package vendingmachine.constants;
 
 public class ExceptionMessages {
