+++ src/main/java/vendingmachine/constants/Integers.java
@@ -1,6 +1,6 @@
 package vendingmachine.constants;
 
-public class Constants {
+public class Integers {
 
     public static int ZERO_COINS = 0;
     public static int NO_CUSTOMER_MONEY_LEFT = 0;
