+++ src/main/java/vendingmachine/constants/Strings.java
@@ -1,0 +1,4 @@
+package vendingmachine.constants;
+
+public class Strings {
+}
